fix generation with multiple multiplicative models
