+++ src/generation/function-generation.jl
@@ -13,101 +13,13 @@ import SpectralFitting:
     modelkind,
     has_closure_params
 
+# include utility functions after struct definition
 include("parsing-utilities.jl")
-
-mutable struct GenerationAggregate{NumType}
-    statements::Vector{Expr}
-    infos::Vector{ModelInfo}
-    closure_params::Vector{Symbol}
-    models::Vector{Type}
-    flux_count::Int
-    maximum_flux_count::Int
-end
-GenerationAggregate(T, infos) =
-    GenerationAggregate{T}(Expr[], [infos], Symbol[], Type[], 0, 0)
-GenerationAggregate(T, infos::AbstractVector) =
-    GenerationAggregate{T}(Expr[], infos, Symbol[], Type[], 0, 0)
-
-push_closure_param!(g::GenerationAggregate, s::Symbol) = push!(g.closure_params, s)
-push_info!(g::GenerationAggregate, s::ModelInfo) = push!(g.infos, s)
-push_statement!(g::GenerationAggregate, s::Expr) = push!(g.statements, s)
-push_model!(g::GenerationAggregate, t::Type) = push!(g.models, t)
 
 function new_closure_param!(ga::GenerationAggregate, p::Symbol)
     param = Base.gensym(p)
     push_closure_param!(ga, param)
     param
-end
-
-get_flux_symbol(i::Int) = Symbol(:flux, i)
-function get_flux_symbol!(g::GenerationAggregate)
-    i = inc_flux!(g)
-    Symbol(:flux, i)
-end
-
-function set_flux!(g::GenerationAggregate, f)
-    g.flux_count = f
-    if g.flux_count > g.maximum_flux_count
-        g.maximum_flux_count = g.flux_count
-    end
-    g.flux_count
-end
-
-inc_flux!(g::GenerationAggregate) = set_flux!(g, g.flux_count + 1)
-dec_flux!(g::GenerationAggregate) = set_flux!(g, g.flux_count - 1)
-
-# model invokation generation
-function add_invoke_statment!(
-    ga::GenerationAggregate{NumType},
-    flux,
-    M::Type{<:AbstractSpectralModel},
-) where {NumType}
-    index = length(ga.models) + 1
-    # infos are in the correct order if indexes this way
-    info = ga.infos[index]
-
-    # aggregate closure parameters
-    closure_params = if has_closure_params(M)
-        map(closure_parameter_symbols(M)) do p
-            new_closure_param!(ga, p)
-        end
-    else
-        ()
-    end
-
-    # get the parameter type
-    T = NumType <: SpectralFitting.FitParam ? NumType.parameters[1] : NumType
-    model_constructor =
-        :($(M.name.wrapper){$(M.parameters[1:end-2]...),$(T),$(M.parameters[end])})
-
-    # assemble the invocation statement
-    s = :(invokemodel!(
-        $flux,
-        energy,
-        $(model_constructor)($(closure_params...), $(info.generated_symbols...)),
-    ))
-    push_model!(ga, M)
-    push_statement!(ga, s)
-end
-
-# don't increment flux for Convolutional models
-function add_invokation!(ga::GenerationAggregate, M, ::Convolutional)
-    flux = get_flux_symbol(ga.flux_count)
-    add_invoke_statment!(ga, flux, M)
-end
-# increment for everything else
-function add_invokation!(ga::GenerationAggregate, M, ::AbstractSpectralModelKind)
-    inc_flux!(ga)
-    flux = get_flux_symbol(ga.flux_count)
-    add_invoke_statment!(ga, flux, M)
-end
-
-function add_flux_resolution!(ga::GenerationAggregate, op::Symbol)
-    fr = get_flux_symbol(ga.flux_count)
-    dec_flux!(ga)
-    fl = get_flux_symbol(ga.flux_count)
-    expr = Expr(:call, op, fl, fr)
-    push_statement!(ga, :(@.($fl = $expr)))
 end
 
 function assemble_closures(ga::GenerationAggregate, model)
@@ -190,28 +102,9 @@ function generated_model_call!(fluxes, energy, model, params)
 end
 
 
-function assemble_aggregate_info(model::Type{<:CompositeModel}, NumType)
-    ga = FunctionGeneration.GenerationAggregate(NumType, getinfo(model))
-    FunctionGeneration.recursive_model_parse(model) do (left, right, op_type)
-        # get operation symbol
-        op = operation_symbol(op_type)
-        if (right !== Nothing)
-            FunctionGeneration.add_invokation!(ga, right, modelkind(right))
-        end
-        if (left !== Nothing)
-            FunctionGeneration.add_invokation!(ga, left, modelkind(left))
-        end
-        if (!isnothing(op))
-            FunctionGeneration.add_flux_resolution!(ga, op)
-        end
-        Nothing
-    end
-    ga
-end
 function assemble_aggregate_info(model::Type{<:AbstractSpectralModel}, NumType)
-    ga = FunctionGeneration.GenerationAggregate(NumType, getinfo(model))
-    flux = FunctionGeneration.get_flux_symbol(FunctionGeneration.inc_flux!(ga))
-    FunctionGeneration.add_invoke_statment!(ga, flux, model)
+    ga = FunctionGeneration.GenerationAggregate(NumType)
+    _addinfoinvoke!(ga, model, :(model))
     ga
 end
